requirements.txt
fastapi
uvicorn[standard]
numpy
opencv-python-headless
insightface==0.7.3
onnxruntime
opencv-python
pymilvus==2.3.3
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
dbutils
mysql-connector-python
geopy==2.4.1
pydantic[email] 
